requirements.txt
pandas==2.2.3
numpy==2.2.1
openai==1.59.3
rich==13.9.4
feedparser==6.0.11
pyshorteners==1.0.1
FlagEmbedding==1.3.3
python-dotenv==1.0.1
torch==2.5.1 
torchaudio==2.5.1
torchvision==0.20.1